requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
supabase==2.3.0
pillow==10.2.0
numpy==1.26.3
httpx==0.26.0
pydantic==2.5.0
python-dotenv==1.0.0
onnxruntime==1.16.3
